Removed unused context objects
