+++ NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Repository/Models/ResearchStudyTeamMember.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
+namespace NIHR.StudyManagement.Infrastructure.Repository.Models
 {
     public partial class ResearchStudyTeamMember : DbEntity
     {
