--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Repository/Models/StudyRegistryContext/PersonDb.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Repository.Models.StudyRegistryContext
-{
-    public partial class PersonDb : DbEntity
-    {
-        public PersonDb()
-        {
-            ResearchInitiativeTeams = new HashSet<ResearchInitiativeTeamMemberDb>();
-        }
-
-        public int Id { get; set; }
-        
-        public string? Firstname { get; set; } = "";
-
-        public string Lastname {get;set;} = "";
-
-        public string PrimaryEmail {get;set;} = "";
-
-        public virtual ICollection<ResearchInitiativeTeamMemberDb> ResearchInitiativeTeams { get; set; }
-    }
-}
