--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Scaffolded02/PersonName.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
-{
-    public partial class PersonName : DbEntity
-    {
-        public int Id { get; set; }
-        public int PersonId { get; set; }
-        public string Family { get; set; } = null!;
-        public string Given { get; set; } = null!;
-
-        public string Email { get; set; } = null!;
-
-        public virtual Person Person { get; set; } = null!;
-    }
-}
