--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Scaffolded02/PersonRole.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
-{
-    public partial class PersonRole : DbEntity
-    {
-        public PersonRole()
-        {
-            ResearchStudyTeamMembers = new HashSet<ResearchStudyTeamMember>();
-        }
-
-        public int Id { get; set; }
-        public string Type { get; set; } = null!;
-        public string? Description { get; set; }
-
-        public virtual ICollection<ResearchStudyTeamMember> ResearchStudyTeamMembers { get; set; }
-    }
-}
