--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Scaffolded02/PersonType.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
-{
-    public partial class PersonType : DbEntity
-    {
-        public PersonType()
-        {
-            People = new HashSet<Person>();
-        }
-
-        public int Id { get; set; }
-        public string Description { get; set; } = null!;
-
-        public virtual ICollection<Person> People { get; set; }
-    }
-}
