--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Scaffolded02/ResearchInitiativeIdentifierType.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
-{
-    public partial class ResearchInitiativeIdentifierType : DbEntity
-    {
-        public ResearchInitiativeIdentifierType()
-        {
-            ResearchInitiativeIdentifiers = new HashSet<ResearchInitiativeIdentifier>();
-        }
-
-        public int Id { get; set; }
-        public string Description { get; set; } = null!;
-
-        public virtual ICollection<ResearchInitiativeIdentifier> ResearchInitiativeIdentifiers { get; set; }
-    }
-}
