--- NIHR.StudyManagement/NIHR.StudyManagement.Infrastructure/Scaffolded02/Researcher.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace NIHR.StudyManagement.Infrastructure.Scaffolded02
-{
-    public partial class Researcher : DbEntity
-    {
-        public Researcher()
-        {
-            ResearchStudyTeamMembers = new HashSet<ResearchStudyTeamMember>();
-        }
-
-        public int Id { get; set; }
-        public int PersonId { get; set; }
-
-        public virtual Person Person { get; set; } = null!;
-        public virtual ICollection<ResearchStudyTeamMember> ResearchStudyTeamMembers { get; set; }
-    }
-}
